pyproject.toml
[tool.poetry]
name = "contacts"
version = "0.0.1"
description = ""
authors = ["Sergey Parhomenko <sergey@parhomenko.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.3"
click = "8.0.4"
google-api-python-client = "2.41.0"
google-auth-oauthlib = "0.5.1"
libgravatar = "1.0.0"
more-itertools = "8.12.0"
phonenumberslite = "8.12.45"
Pillow = "9.0.1"
pyobjc-framework-Contacts = "8.4.1"
PyYAML = "6.0"
requests = "2.27.1"
Telethon = "1.24.0"
waiting = "1.4.1"

[tool.poetry.dev-dependencies]
black = "22.1.0"
flake8 = "4.0.1"
mypy = "0.941"
types-PyYAML = "6.0.5"
types-requests = "2.27.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 200
target-version = ["py310"]

[tool.isort]
line_length = 200
profile = "black"
py_version = 310
